requirements.txt
fastapi==0.111.0
cassandra-driver==3.29.1
uvicorn==0.30.1
python-dotenv=1.0.1
PyJWT=2.8.0
pydantic=2.8.2
pytz=2024.1
requests=2.32.3